requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.1
click==8.3.1
colorama==0.4.6
distro==1.9.0
docopt==0.6.2
durationpy==0.10
edge-tts==7.2.7
filelock==3.24.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
num2words==0.5.14
numpy==2.4.2
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.24.2
openai==2.21.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pillow==12.1.1
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.5
PyAudio @ file:///C:/Users/matve/Downloads/PyAudio-0.2.14-cp311-cp311-win_amd64.whl#sha256=bbeb01d36a2f472ae5ee5e1451cacc42112986abe622f735bb870a5db77cf903
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
fastapi>=0.115.0
onnx-asr[gpu]>=0.1.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pystray==0.19.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.3
shellingham==1.5.4
silero-vad
silero==0.5.5
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
srt==3.5.3
sympy==1.14.0
tabulate==0.9.0
tavily-python==0.7.21
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
ttkbootstrap==1.20.1
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
yarl==1.22.0
zipp==3.23.0
